labs/week1/practice/cuongld/chatbot/requirements.txt
streamlit>=1.33.0
python-dotenv>=1.0.1
langgraph>=0.6.6
langgraph-checkpoint-sqlite>=2.0.0
langchain-openai>=0.3.0
aiosqlite>=0.21.0
fpdf2>=2.7.9